requirements.txt
flask
flask_socketio
eventlet

opuslib
pyaudio

pyyaml
